Plugged in grid function, created function to generate URLs from gridpoints
Connected to pgsql database (using Ben's example db for now)

Built a function to generate URLs in the correct format from the gridpoints table in the database. The format appears correct, but OTP does not recognize the start points. Possible that the OSM data we're using does not cover a wide enough area, so the points I've tried fall outside of it?

Plugged in generateGrid sql query